data-engineering/requirements.txt
requests==2.32.*
kafka-python==2.0.*